refactor: image providers into subfolder
